Moved `ChoosePropertyPage` outside `ReorderUnitsPage`
